+++ internal/image/auto_backend.go
@@ -68,7 +68,7 @@ func (b *AutoImageBackend) RemoveImage(ctx context.Context, req ImageRemoveReque
 
 func (b *AutoImageBackend) backend(ctx context.Context) (ImageBackend, error) {
 	if b == nil {
-		return nil, imageBackendOpError{Op: "select image backend", Err: fmt.Errorf("auto image backend is required")}
+		return nil, BackendOpError{Op: "select image backend", Err: fmt.Errorf("auto image backend is required")}
 	}
 	mode := strings.ToLower(strings.TrimSpace(b.mode))
 	if mode == "" {
@@ -89,13 +89,13 @@ func (b *AutoImageBackend) backend(ctx context.Context) (ImageBackend, error) {
 		b.lastSelection = appconfig.ImageStoreModeOCI
 		return b.requireBackend(b.oci, appconfig.ImageStoreModeOCI)
 	default:
-		return nil, imageBackendOpError{Op: "select image backend", Err: fmt.Errorf("unsupported image store mode %q", b.mode)}
+		return nil, BackendOpError{Op: "select image backend", Err: fmt.Errorf("unsupported image store mode %q", b.mode)}
 	}
 }
 
 func (b *AutoImageBackend) requireBackend(backend ImageBackend, name string) (ImageBackend, error) {
 	if backend == nil {
-		return nil, imageBackendOpError{Op: "select image backend", Err: fmt.Errorf("%s image backend is required", name)}
+		return nil, BackendOpError{Op: "select image backend", Err: fmt.Errorf("%s image backend is required", name)}
 	}
 	return backend, nil
 }
